Google-Meet-Bot/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==1.26.4
openai==1.99.9
outcome==1.3.0.post0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.1
scipy==1.11.4
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
